requirements.txt
fastapi
flask
uvicorn
pydantic
pyjwt
python-dotenv
mysql-connector-python
bcrypt
flask_cors

